moved create bracket modal to its own component
